Correctly call fixture test_start for each test
Currently if there is an exception raised from one of the fixture
test_start, the test before and other tests method are not called.
In addition, this fixture test_start will not be called again until all
the fixtures test_start have been called once without exceptions.
This leads to funky results...