Modifying the fixture test_start unit test to fix active fixture order
Previous implementation had parametrization on the fixture raising
ZeroDivisionError, this caused the second fixture called first staring
on the second tests.

This is because only the second fixture stay active.

I am not sure if we should fix this since we are not always honoring
the order of the fixture in the test method argument.