Add loglinkconverter plugin
When tests are run on a continuous integration system, this plugin
converts the local filesystem links to log files given in the test
session summary into web links into the artifact store for the job
which ran the tests. As a result, the test summary sections for
each failing test will contain a clickable link to the log file
for that test, making it easier to quickly diagnose a failing test.